Add the Nintendo 3DS port
The whole game rather than a reimplementation: the same physics, the same
configuration language and the same network protocol, so it plays on the
public servers alongside desktop clients.

Thirty-two files in src/ carry the port's changes and the rest of it lives
under platform/3ds: a compatibility layer over citro3d standing in for the
OpenGL the PICA200 does not have, a dual screen split for the menus and the
arena map, 3DS input, and a runtime that reports where it died rather than
dropping to the Homebrew Launcher in silence.

protobuf, libxml2, FTGL and the boost headers are not in devkitPro's
portlibs, so they are vendored under platform/3ds/vendor and cross built by
build-dependencies.sh. That directory needs its own ignore list: the one at
the top of the repository is written for the autotools build that lives
there, and a bare "Makefile" rule would have quietly dropped the 3DS
makefile and left the port unbuildable for anyone who cloned it.

Building is the three shell scripts next to that makefile, which is also
what the workflow runs, so releases are not built by a path nobody
exercises. Pushing a tag beginning with v builds through the same workflow
and publishes the 3dsx and the CIA against it.

Not carried over from the development repository: build output, the RomFS
image and the protobuf sources, all of which are generated; the CMake trees
for the vendored dependencies, which were full of absolute paths from one
machine; and the patch series the port was originally assembled from, which
this history now replaces.

+++ src/network/nProtoBuf.cpp
@@ -710,9 +710,9 @@ void nProtoBufDescriptorBase::StreamFromDefault( nStreamMessage & in, nProtoBuf 
         {
         case FieldDescriptor::CPPTYPE_INT32:
         {
-            int32 value;
+            int value;
             in >> value;
-            reflection->REFL_SET( SetInt32, &out, field, value );
+            reflection->REFL_SET( SetInt32, &out, field, static_cast< int32 >( value ) );
         }
         break;
         case FieldDescriptor::CPPTYPE_UINT32:
@@ -863,7 +863,7 @@ void nProtoBufDescriptorBase::StreamToDefault( nProtoBuf const & in, nStreamMess
         switch( field->cpp_type() )
         {
         case FieldDescriptor::CPPTYPE_INT32:
-            out << reflection->REFL_GET( GetInt32, in, field );
+            out << static_cast< int >( reflection->REFL_GET( GetInt32, in, field ) );
             break;
         case FieldDescriptor::CPPTYPE_UINT32:
             out.Write( reflection->REFL_GET( GetUInt32, in, field ) );
